Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -69,6 +69,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -81,8 +84,12 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title', 'body'));
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
 
+        $question->update($request->only('title', 'body'));
+        
         return redirect('/questions')->with('success', "Your question has been updated");
     }
 
@@ -93,7 +100,10 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {       
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success', "Your question has been deleted");
     }
